requirements.txt
streamlit
beautifulsoup4
requests
python-dotenv
langchain
langchain-community
openai